Make DeviceActions suspend until complete
Since they are suspend functions, they should not launch asynchronous
jobs to activate / deactivate devices; they can just change context as
needed and return when complete. This simplifies our tests a bit since
we know the handles will have changed states after the action is
complete.

This exposes a race condition in shutdown of AVDs: the onKill action
of the FakeEmulatorConsole closes the virtual AVD, sometimes resulting
in the EmulatorConsole being closed before the "OK" response to "kill"
makes it back to the client, resulting in a ChannelClosedException. This
is benign and we just need to catch it.

Bug: 241591222
Test: updated
Change-Id: Ia6c61574856d8f1a89349f80dd7bbbb73128f4ed