Adapt to dynamic power track feature flag
On studio/idea side, the feature flag for displaying
the power rail and battery counter tracks was made to
be an enum/integer to be more dynamic (the positive integers
are mapped to enums on studio-side). This now allows us
to disable the viewing of the power and battery tracks,
but allows us to add different visualizations via non-zero
positive integers mapping to a different visualization query.

This change adapts the trace processor daemon and related
protos to accept and utilize the aforementioned passed in
feature flag.

Bug: 254116554
Test: existing
Change-Id: I62a95eeb5e38472e8aa00adb23bcf2c8e06a660f