Stop uploading coverage artifacts to GCS
Also delete the commented out HTML report generation.

With the deprecation of the old dashboard/pipeline, this is no longer
useful. It's only remaining use was for the fresh builds identifier
cronjob to find builds ready to upload. That job has been permanently
shutdown now.

Artifacts currently in GCS will get cleaned up naturally due to the
"delete after 6 months" policy on the bucket. If we want to do it
earlier then we can execute a manual mass deletion.

Fixes: 265360442
Fixes: 265360419
Test: N/A
Change-Id: I55d3d6af42ccd7904265f6265cccf5142d8cd8c6