Create FakeClock utility class for testing
Some of our code uses the Java8 Time APIs. Unfortunately, Java
does not provide a fake implementation of Clock() and leaves it
up to developers to write their own. We have a couple but they
are spread out all over the place. This will ensure that we have
one more. Maybe also we can replace the others.

Bug: N/A
Test: FakeClockTest.kt
Change-Id: I4b62ab2b1957bed8ae952eaed7cf96f223607652