Fix many CleanupDetector false positives, and improve flow analyzer
This CL fixes many false positives in the CleanupDetector found by
an audit of the Android framework and g3 codebase:

267308328: false positive for resource returned in an array
267462840: false positive for a wrong ofInt() method
267439692: false positive for AssetFileDescriptor
267597964: false positive for resource returned in ?: operator
267592743: false positive when calling this() constructor from resource code
267743132: false positive for resource assigned to a variable after type casting
267743930: false positive for resource created right before try-with-resources
267597120: false positive with ignored method reference

Most of these were relatively straightforward, but the last one, method
references, required a bit of work. The key problem is that we may
actually reach a method reference for the target method, but we can't
yet conclude that you're actually invoking that reference, so we need to
start tracking it separately. To handle this, not just in the
CleanupDetector, but all places where we're doing similar flow analysis,
I introduced a new subclass of the DataFlowAnalyzer intended for
analyzers which are looking for a target method. I then retrofitted
various users of the data flow analyzer to this new subclass instead.
(This CL also updates the API documentation a bit about this.)

Fixes: 267308328
Fixes: 267462840
Fixes: 267439692
Fixes: 267597964
Fixes: 267592743
Fixes: 267743132
Fixes: 267743930
Fixes: 267597120
Test: Unit tests included
Change-Id: Ie86cfcff0991ad9a39c7c9b3cf9c18058928bd91