adblib: Add `device` property to `JdwpSession`
Adding a `device: ConnectedDevice` property to `JdwpSession`
and `SharedJdwpSession` so consumer have access to the
device the session is connected to makes sense for scoping,
extensions, etc.

In particular, this change enables a follow-up change to use
the device cache to store custom property.

Most of the code are "mechanical" changes to ensure
`ConnectedDevice` is passed to the constructor of
`JdwpSessionImpl`.

Bug: 241411829
Bug: 188067646
Test: Existing
Change-Id: Ie0ef94c49f3e974a9335786ba2b49d0333b1c056