Remove module to project dependency substitution
From Gradle 8.0 dependency substitution in before resolve is not
permitted. Currently there is automatic project to module dependency substitution
when a module is requested from a project, which leads to a Gradle
issue. However, removing this substition breaks implicit substitution
for android tests, so this will need to be addressed with a more
specific check to provide users with an alternative configuration.

The ProjectSubstitutionDependenciesTest snapshot test appears to show
user applied dependency substitution was not being applied correctly
before this change. The snapshot now shows the expected substituted
dependencies.

Bug: 253687111
Test: ProjectSubstitutionDependenciesTest
Change-Id: Ib0c0ffa021a3b2ed61705f685ffac06b710fb5eb