Read SDK extension levels from device
When building the AndroidVersion from device properties, we need to
consider the "build.version.extensions.*" properties to get the
extension level.

Note that these are not "ro.*" properties (because they are reset
during userspace reboot), so we need to start passing all properties to
readCommonProperties(), not just the readonly ones.

Fixes: 264869521
Test: added
Change-Id: If4f494a33b48a306e77a80e8a5286cc95eb129f7