Set up dependencies in settings for integration tests
In order to run integration tests with project isolation, we need to
get rid of violations like using allProjects {} to set up repositories
for dependency resolution.

Apart from that, as we switch to use dependencyResolutionManagement
block in settings.gradle, we shouldn't set repo again in the projects.
This CL addresses some of them but not all. When all the clean up is
finished, RepositoriesMode.FAIL_ON_PROJECT_REPOS will be used as
default. However, if there is a need to prefer setting repo in project,
we could introduce a flag to change RepositoriesMode.

Bug: 270584807
Test: existing
Change-Id: I6e0e2857b1403ef541318f78cc2af770355e3570