Disable obfuscation and optimization for desugar library apis
Shrinking is needed for desugar library apis in release build because we
need the apk size to be small. However, the obfuscation and optimization
need to be disabled in release build, otherwise the app would just crash
saying cannot find specifc desugared api.

Prior to this change, the keep rules for desugar library have been
changed since we moved to use trace reference tool and keep rules
started to include configuration of "allowobfuscation". This doesn't
cause problem in non-minified build becasue we also add "-dontobfuscate"
in non minified release build. However, it causes a problem in minified
relese build where we don't pass "-dontobfuscate". Note, "-dontobfuscate"
can let "allowobfuscation" to be ignored.

To fix this, first we need to ensure the keep rules are generated
same as before by removing "allowobfuscation". Second, as suggested
by r8 team, we should also add "-dontoptimize" along with
"-dontobfuscate" to disable inlining. Third, as the keep rules are not
comprehensive to cover all classes which shouldn't be obfuscated, we
need to add "-dontobfuscate" as a global configuration in all release
build.

Bug: 266687543
Test: CoreLibraryDesugarTest + CoreLibraryDesugarConnectedTest
Change-Id: I6290c3f7f041b9b8330e69e4e5dbf5dbb187d550