adblib: FakeAdbServer sends `APNM` after `HELO`
Sending `APNM` ddms command just after sending
the reply to the `HELO` ddms command makes
FakeAdbServer behave a little bit closer to the
behavior of a real device.

This change allows improving code coverage of
adblib tests, as the code to handle `APNM`
commands from a JDWP session was untested.

Bug: 188067646
Test: Updated existing tests
Change-Id: I7e39e8f714d04f0a328f473de4b691a2585a1cf1