adblib: Fix WAIT packet behavior
A recent change in the JDWP process properties collector
allowed process collections to stop as soon as a ddms
`WAIT` packet is received.

Unfortunately, this change (in some cases) breaks
the Android Studio debugger (specifically Instrumented
test debugger), because the Android debugger gets
into a "hung" state when a JDWP session is closed
between the Android debugger sending the WAIT
packet and the "external" debugger (Android Studio)
sending the first JDWP command packet. This is
because the Android debugger waits for that initial
JDWP packet to continue.

Bug: 271466829
Test: Existing
Change-Id: I399ec064ab084ff4da1ba9cc07336515e0813522