[toml] Fix for invalid offsets in warnings
When we start parsing a Toml value, it's fairly natural to know the
start offset (we are at it, after all) but not to know its end
offset (we need to do some parsing).  That implies that in the
presence of sufficiently complex invalid Toml, it is easy to get to
the end of a stream without having ever detected the end condition for
a value, which means that `endOffset` will still be in its
uninitialized state; if we then try to warn, we need the uninitialized
state to be valid for the warning.

The natural default for `endOffset` of a new TomlValue is therefore
probably not -1, but rather the endOffset of its parent: with the idea
that we can successively refine the endOffsets of child values as we
discover them, but if we never discover new information about the
current value, then we will end up at end of file (which is the
already-known endOffset of the document root).

Bug: 271454831
Test: added
Change-Id: I6f698ea02b0995f17a4bc30c1776e4a3f265691a